more adjustments for worker jobs dispatch + refactor shared function packages + remove py2 references
