+++ requirements-dev.txt
@@ -5,7 +5,6 @@ astroid>=2.4
 bandit
 bump2version
 codacy-coverage
-contextlib2; python_version < "3"
 coverage
 doc8>=0.8.1
 flake8
