+++ requirements.txt
@@ -1,5 +1,4 @@
 alembic
-backports.tempfile; python_version < "3"
 # AWS support (S3 buckets)
 boto3
 # celery 5 to be released in near future
@@ -11,12 +10,10 @@ cornice
 cornice_swagger>=0.7.0
 # cwltool Python 2 support dropped on 2.x
 # enforce Python 3 with version >=2 to update existing 1.x installations
-cwltool>=1.0.20180820141117,<2; python_version < "3"
 # FIXME: avoid error invalid get_listing import location,
 #        remove properly as per https://github.com/crim-ca/weaver/issues/154
 cwltool>=2,<=3.0.20200324120055; python_version >= "3"
 duration
-git+https://github.com/ESGF/esgf-compute-api.git@v2.1.0#egg=cwt ; python_version < "3"
 git+https://github.com/ESGF/esgf-compute-api.git@v2.3.7#egg=esgf-compute-api ; python_version > "3"
 flufl.enum; python_version < "3"
 # gunicorn >20 breaks some config.ini loading parameters (paste)
@@ -37,8 +34,8 @@ pytz
 # no viable pywps version with Python>3.5 dependencies
 # use '4.2.4' plus a few commits that provide fix, but not yet released
 # FIXME: https://github.com/geopython/pywps/issues/568
-git+https://github.com/fmigneault/pywps.git@799fb14f31533630ce08c171f844294730861b1a#egg=pywps ; python_version >= "3"
-pywps==4.2.4; python_version < "3"
+## git+https://github.com/fmigneault/pywps.git@799fb14f31533630ce08c171f844294730861b1a#egg=pywps ; python_version >= "3"
+git+https://github.com/geopython/pywps@pywps-4.4#egg=pywps ; python_version >= "3"
 pyyaml>=5.2
 requests
 requests_file
@@ -47,9 +44,7 @@ requests_file
 # (https://bitbucket.org/ruamel/yaml/issues/261/error-while-installing-ruamelyaml-setuppy)
 # let cwltool fully define the version if python 2 is used
 ruamel.yaml>=0.15.78,<=0.16.5; python_version >= "3"
-ruamel.yaml<=0.16.5; python_version < "3"
 shapely
-six
 simplejson
 urlmatch
 xmltodict
