+++ tests/test_utils.py
@@ -1,5 +1,6 @@
 # pylint: disable=C0103,invalid-name
 
+import contextlib
 import inspect
 import json
 import os
@@ -22,9 +23,8 @@ from pyramid.httpexceptions import (
 from pywps.response.status import WPS_STATUS
 from requests import Response
 from requests.exceptions import HTTPError as RequestsHTTPError
-from six.moves.urllib.parse import urlparse
+from urllib.parse import urlparse
 
-from tests.compat import contextlib
 from tests.utils import mocked_aws_credentials, mocked_aws_s3, mocked_aws_s3_bucket_test_file, mocked_file_response
 from weaver import status, utils
 from weaver.utils import _NullType  # noqa: W0212
