+++ weaver/processes/builtin/utils.py
@@ -2,14 +2,12 @@ import os
 from typing import Any
 from urllib.parse import urlparse
 
-import six
-
 from weaver.formats import CONTENT_TYPE_APP_NETCDF, get_extension
 
 
-def _is_netcdf_url(url):
+def is_netcdf_url(url):
     # type: (Any) -> bool
-    if not isinstance(url, six.string_types):
+    if not isinstance(url, str):
         return False
     if urlparse(url).scheme == "":
         return False
