+++ weaver/processes/wps3_process.py
@@ -43,9 +43,11 @@ from weaver.wps_restapi.swagger_definitions import (
 )
 
 if TYPE_CHECKING:
-    from weaver.typedefs import JSON, UpdateStatusPartialFunction   # noqa: F401
-    from typing import Union, AnyStr                                # noqa: F401
-    from pywps.app import WPSRequest                                # noqa: F401
+    from typing import Union
+
+    from pywps.app import WPSRequest
+
+    from weaver.typedefs import JSON, UpdateStatusPartialFunction
 
 LOGGER = logging.getLogger(__name__)
 
@@ -63,7 +65,7 @@ class Wps3Process(WpsProcessInterface):
     def __init__(self,
                  step_payload,      # type: JSON
                  joborder,          # type: JSON
-                 process,           # type: AnyStr
+                 process,           # type: str
                  request,           # type: WPSRequest
                  update_status,     # type: UpdateStatusPartialFunction
                  ):
